refactor main module into own file, add relative imports, app functions for plotting interface
